Model

+++ ui-server/model/writer.py
@@ -1,68 +1,48 @@
 import threading
 import time
+import datetime
 import csv
-
+WRITE_MODE = 'w' # 'w' for overwrite, 'a' for append
+#   -----Description-----
+#   This is a threaded class that writes data
+#   to a CSV file every self.writeInterval seconds.
+#   ---------------------
 class writer(threading.Thread):
-    def __init__(self, threadID, fieldnames, filename, sensors):
+    # Specify a time delay, a list of CSV column header names, the CSV filename, and the sensorHelper object
+    def __init__(self, writeInterval, fieldnames, filename, sHelper):
         threading.Thread.__init__(self)
 
-        self.THREAD_SLEEP_TIME = .1
+        self.writeInterval = writeInterval
 
-        self.threadID = threadID
         self.fieldnames = fieldnames
+        self.fieldnames.append("Timestamp") # Add a field for timestamps so clients don't have to worry about it
+
         self.filename = filename
-        # self.counter = 0
-        self.sensors = sensors
-        self.dataList = []
-
-        self.__updateList()
-
-        with open(self.filename, mode='w', newline='\n') as self.csvfile:
-            self.writer = csv.DictWriter(self.csvfile, fieldnames=self.fieldnames)
-            self.writer.writeheader()
+        self.sHelper = sHelper
+        if WRITE_MODE == 'w': # If in overwrite mode, rewrite headers (which deletes all file data)
+            with open(self.filename, mode='w', newline='\n') as self.csvfile:
+                self.writer = csv.DictWriter(self.csvfile, fieldnames=self.fieldnames) # Lets us write each column in a single row separately
+                self.writer.writeheader()
     def run(self):
         while True:
-
+            timestampValue = datetime.datetime.now() # Gets the current time
+            # This isn't very efficient (opens & closes file every loop), but it works for our purposes
             with open(self.filename, mode='a', newline='\n') as self.csvfile:
+                ss = self.sHelper.getSS()
                 self.writer = csv.DictWriter(self.csvfile, fieldnames=self.fieldnames)
-                self.__updateList()
-
-                # for i in range(len(self.fieldnames)):
-                #     self.writer.writerow({self.fieldnames[i]: self.dataList[i]})
-
-                self.writer.writerow({self.fieldnames[0]: self.dataList[0], self.fieldnames[1]: self.dataList[1],
-                                      self.fieldnames[2]: self.dataList[2], self.fieldnames[3]: self.dataList[3],
-                                      self.fieldnames[4]: self.dataList[4], self.fieldnames[5]: self.dataList[5],
-                                      self.fieldnames[6]: self.dataList[6], self.fieldnames[7]: self.dataList[7],
-                                      self.fieldnames[8]: self.dataList[8], self.fieldnames[9]: self.dataList[9],
-                                      self.fieldnames[10]: self.dataList[10], self.fieldnames[11]: self.dataList[11],
-                                      self.fieldnames[12]: self.dataList[12], self.fieldnames[13]: self.dataList[13],
-                                      self.fieldnames[14]: self.dataList[14], self.fieldnames[15]: self.dataList[15],
-                                      self.fieldnames[16]: self.dataList[16], self.fieldnames[17]: self.dataList[17],
-                                      self.fieldnames[18]: self.dataList[18], self.fieldnames[19]: self.dataList[19],
-                                      self.fieldnames[20]: self.dataList[20]
-
-                                      })
-
-                # print(self.threadID, self.counter)
-                # self.counter+=1
-                time.sleep(self.THREAD_SLEEP_TIME)
-
-    def __updateList(self):
-        self.dataList.clear()
-
-        self.dataList.append(self.sensors[0].getValue().pop())
-        self.dataList.append(self.sensors[0].getValue().pop())
-
-        self.dataList.append(self.sensors[1].getValue().pop())
-        self.dataList.append(self.sensors[2].getValue().pop())
-        self.dataList.append(self.sensors[3].getValue().pop())
-
-        for i in range(16):
-            self.dataList.append(self.sensors[4].getValue().pop())
-
-
-
+                self.writer.writerow({
+                        self.fieldnames[0]: ss[0], self.fieldnames[1]: ss[1], # Writes each ss list value in the appropriate self.fieldnames column
+                        self.fieldnames[2]: ss[2], self.fieldnames[3]: ss[3],
+                        self.fieldnames[4]: ss[4], self.fieldnames[5]: ss[5],
+                        self.fieldnames[6]: ss[6], self.fieldnames[7]: ss[7],
+                        self.fieldnames[8]: ss[8], self.fieldnames[9]: ss[9],
+                        self.fieldnames[10]: ss[10], self.fieldnames[11]: ss[11],
+                        self.fieldnames[12]: ss[12], self.fieldnames[13]: ss[13],
+                        self.fieldnames[14]: ss[14], self.fieldnames[15]: ss[15],
+                        self.fieldnames[16]: ss[16], self.fieldnames[17]: ss[17],
+                        self.fieldnames[18]: ss[18], self.fieldnames[19]: ss[19],
+                        self.fieldnames[20]: ss[20], self.fieldnames[21]: timestampValue})
+                time.sleep(self.writeInterval)
